requirements.txt
tensorflow-gpu==1.14
joblib
sklearn
keras
pandas
opencv-python==4.2.0.32
matplotlib
ipython
imageio==2.6
scikit-image
tensorflow-determinism
